web/requirements.txt
flask==2.2.2
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
flask-migrate
faker
psycopg2